pyproject.toml
[project]
name = "hakymi-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "deepagents>=0.4.12",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.3",
    "pygraphviz",
    "python-dotenv>=1.2.2",
    "tavily-python>=0.7.23",
]
